Refactor fragment replacement and navigation #139
From now on, the visibility of UI for specific fragments will be handled
when the corresponding fragment resumes or stops. Also, the menu drawer
and selection will be decided when resuming each corresponding fragment.

Activity callbacks have also been added for scalability purposes in case
we plan to reuse some of these fragments in other activities.